- Added a temporary band-aid to prevent executing the script of non-arrows when they are equipped on the arrow-slot (until the actual cause of why pc_unequipitem is not working correctly is discovered/fixed)
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@7559 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/pc.c
@@ -57,7 +57,7 @@ struct fame_list smith_fame_list[MAX_FAME_LIST];
 struct fame_list chemist_fame_list[MAX_FAME_LIST];
 struct fame_list taekwon_fame_list[MAX_FAME_LIST];
 
-static unsigned int equip_pos[11]={0x0080,0x0008,0x0040,0x0004,0x0001,0x0200,0x0100,0x0010,0x0020,0x0002,0x8000};
+static unsigned short equip_pos[11]={0x0080,0x0008,0x0040,0x0004,0x0001,0x0200,0x0100,0x0010,0x0020,0x0002,0x8000};
 
 static struct gm_account *gm_account = NULL;
 static int GM_num = 0;
